[TEST] [REORG] Consolidate pyproject.toml tests and improve parsing coverage
This change improves the `pyproject.toml` parser in `gptscan.py` to support
`gui-scripts` sections and quoted keys (with spaces). It also reorganizes
the test suite by consolidating seven fragmented test files into a
single comprehensive module `tests/test_pyproject_parsing.py`.
