--- tests/test_pyproject_robustness.py
@@ -1,96 +1,0 @@
-import pytest
-from gptscan import unpack_content
-
-def test_pyproject_pdm_nested_scripts():
-    content = b"""
-[tool.pdm.scripts.test]
-cmd = "pytest tests"
-help = "Run tests"
-
-[tool.pdm.scripts.post_install]
-shell = "echo done"
-
-[tool.pdm.scripts.multi]
-composite = ["lint", "test"]
-"""
-    snippets = list(unpack_content("pyproject.toml", content))
-    scripts = {s[0]: s[1].decode() for s in snippets}
-
-    assert "pyproject.toml [Script: test]" in scripts
-    assert scripts["pyproject.toml [Script: test]"] == "pytest tests"
-
-    assert "pyproject.toml [Script: post_install]" in scripts
-    assert scripts["pyproject.toml [Script: post_install]"] == "echo done"
-
-    assert "pyproject.toml [Script: multi (1)]" in scripts
-    assert scripts["pyproject.toml [Script: multi (1)]"] == "lint"
-    assert "pyproject.toml [Script: multi (2)]" in scripts
-    assert scripts["pyproject.toml [Script: multi (2)]"] == "test"
-
-def test_pyproject_hatch_scripts():
-    content = b"""
-[tool.hatch.scripts]
-test = "pytest"
-nested = { cmd = "echo hatch" }
-"""
-    snippets = list(unpack_content("pyproject.toml", content))
-    scripts = {s[0]: s[1].decode() for s in snippets}
-
-    assert "pyproject.toml [Script: test]" in scripts
-    assert scripts["pyproject.toml [Script: test]"] == "pytest"
-
-    assert "pyproject.toml [Script: nested]" in scripts
-    assert scripts["pyproject.toml [Script: nested]"] == "echo hatch"
-
-def test_pyproject_mixed_scripts():
-    content = b"""
-[project.scripts]
-cli = "pkg:main"
-
-[tool.pdm.scripts]
-pdm-flat = "echo flat"
-pdm-table = { shell = "echo table" }
-
-[tool.pdm.scripts.pdm-nested]
-cmd = "echo nested"
-"""
-    snippets = list(unpack_content("pyproject.toml", content))
-    scripts = {s[0]: s[1].decode() for s in snippets}
-
-    assert "pyproject.toml [Script: cli]" in scripts
-    assert scripts["pyproject.toml [Script: cli]"] == "pkg:main"
-    assert "pyproject.toml [Script: pdm-flat]" in scripts
-    assert scripts["pyproject.toml [Script: pdm-flat]"] == "echo flat"
-    assert "pyproject.toml [Script: pdm-table]" in scripts
-    assert scripts["pyproject.toml [Script: pdm-table]"] == "echo table"
-    assert "pyproject.toml [Script: pdm-nested]" in scripts
-    assert scripts["pyproject.toml [Script: pdm-nested]"] == "echo nested"
-
-def test_pyproject_case_insensitivity():
-    content = b"""
-[TOOL.pdm.SCRIPTS.UPPER]
-CMD = "echo upper"
-"""
-    snippets = list(unpack_content("pyproject.toml", content))
-    scripts = {s[0]: s[1].decode() for s in snippets}
-
-    assert "pyproject.toml [Script: UPPER]" in scripts
-    assert scripts["pyproject.toml [Script: UPPER]"] == "echo upper"
-
-def test_pyproject_header_robustness():
-    """Verify that pyproject.toml sections with trailing comments and whitespace are recognized."""
-    content = b"""
-[tool.pdm.scripts] # comment
-test = "echo hello"
-
-[ tool.pdm.scripts.nested ] # another comment
-cmd = "pytest"
-"""
-    results = list(unpack_content("pyproject.toml", content))
-    scripts = {s[0]: s[1].decode() for s in results}
-
-    assert len(results) == 2
-    assert "pyproject.toml [Script: test]" in scripts
-    assert scripts["pyproject.toml [Script: test]"] == "echo hello"
-    assert "pyproject.toml [Script: nested]" in scripts
-    assert scripts["pyproject.toml [Script: nested]"] == "pytest"
